+++ lib/src/generator/transfer_bean/copy_with_extension_builder.dart
@@ -1,4 +1,5 @@
-import 'package:cl_datahub_common/src/generator/transfer_field.dart';
+
+import 'transfer_field.dart';
 
 class CopyWithExtensionBuilder {
   final String transferClass;